fix: review conversations fixes
